[fix] Stopped treating @ and ' as word-characters (4h)
They were treated as word characters so that we could substitute invocations like "@me" and contractions like "ain't" before they were tokenizes, just using a Hash map.

By not treating them as word characters, to perform those same substitutions in the tokenizer, we had to perform substitutions using strings of tokens — and to keep from losing the performance benefit of using a Hash map, we had to introduce a Trie to keep track of candidate substitutions.

Aside from treating @ and ' as punctuation characters when they can't be matched, this fix has several other advantages:
  1. Matching contractions and invocations no longer need separate code paths
  2. Contractions can be expanded recursively — "couldn't've" can be first expanded to "could not've" and then to "could not have"
  3. Shorthand like "w/o" can now be expanded to "without" without having to treat / as a word-character
  4. Invocations can now be multi-word like "OK Google"